Refactoring for in preparation for new features
- Created a new `uapi` directory to keep track of rust-friendly apis and
  usage of FFI wrappers and types.
- Split both `linux` and `uapi` into host and guest modules for easier
  management.
- Applied changes to existing imports to reflect these changes.
