+++ src/firmware/linux/guest/mod.rs
@@ -1,0 +1,3 @@
+//! Guest FFI Wrappers for C Kernel APIs
+pub(crate) mod ioctl;
+pub mod types;
