+++ src/firmware/linux/host/mod.rs
@@ -1,0 +1,3 @@
+//! Host FFI Wrappers for C Kernel APIs
+pub(crate) mod ioctl;
+pub mod types;
