+++ src/firmware/uapi/guest/mod.rs
@@ -1,0 +1,3 @@
+//! Guest Owner Rust-friendly API interfaces.
+
+mod types;
