+++ src/firmware/uapi/mod.rs
@@ -1,0 +1,3 @@
+//! Rust-friendly API wrappers to communicate with the FFI functions.
+pub mod guest;
+pub mod host;
